Abstract book layout
