+++ src/pages/Books/Books.module.scss
@@ -1,9 +1,3 @@
-.bookSection {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
 .bookGrid {
   display: grid;
   grid-template-columns: 1fr 1fr;
@@ -13,63 +7,4 @@
     grid-template-columns: repeat(auto-fill, minmax(180px, 1fr));
     gap: 3rem;
   }
-}
-
-.currentReadingTitle {
-  font-size: 1.25rem;
-  text-decoration: underline;
-  text-decoration-color: $primary-color;
-  text-decoration-thickness: 0.4rem;
-  text-decoration-skip-ink: none;
-  text-underline-offset: -4px;
-
-  @media (min-width: $screen-sm-min) {
-    font-size: 1.5rem;
-  }
-}
-
-.searchSection {
-  margin-top: 6rem;
-}
-
-.searchHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 3rem;
-}
-
-.searchInput {
-  display: none;
-  padding: 0.3rem 1rem;
-  border: 1px solid #ddd;
-  border-radius: 24px;
-  width: 300px;
-
-  &:focus {
-    outline: none;
-  }
-
-  @media (min-width: $screen-md-min) {
-    display: block;
-  }
-}
-
-.searchIcon {
-  display: flex;
-  justify-content: center;
-  position: relative;
-  background-color: transparent;
-  cursor: pointer;
-  border: none;
-
-  @media (min-width: $screen-md-min) {
-    display: none;
-  }
-}
-
-.searchInputOpen {
-  display: block;
-  position: relative;
-  width: 180px;
 }
